owners

+++ logs/views.py
@@ -1,6 +1,8 @@
+from django import http
 from django.shortcuts import render, redirect
 from .forms import *
 from .models import Topic
+from django.contrib.auth.decorators import login_required
 
 
 def index(request):
@@ -20,6 +22,9 @@ def topic(request, topic_id):
     obj = Topic.objects.get(id=topic_id)
 
     if request.method == "POST" and request.POST.get("_method") == "DELETE":
+        if request.user != obj.owner:
+            raise http.Http404("You are not allowed.")
+
         obj.delete()
         return render(request, 'logs/topics.html', {"topics": Topic.objects.all()})
     else:
@@ -28,6 +33,7 @@ def topic(request, topic_id):
         return render(request, 'logs/topic.html', context)
 
 
+@login_required
 def new_topic(request):
     """Add a new topic."""
     if request.method != 'POST':
@@ -37,7 +43,9 @@ def new_topic(request):
         # POST data submitted; process data.
         form = TopicForm(data=request.POST)
         if form.is_valid():
-            form.save()
+            new_topic = form.save(commit=False)
+            new_topic.owner = request.user
+            new_topic.save()
             return redirect('logs:topics')
 
     # Display a blank or invalid form.
@@ -45,6 +53,7 @@ def new_topic(request):
     return render(request, 'logs/new_topic.html', context)
 
 
+@login_required
 def new_entry(request, topic_id):
     """Add a new entry for a particular topic."""
     topic = Topic.objects.get(id=topic_id)
@@ -66,10 +75,14 @@ def new_entry(request, topic_id):
     return render(request, 'logs/new_entry.html', context)
 
 
+@login_required
 def edit_entry(request, entry_id):
     """Edit an existing entry."""
     entry = Entry.objects.get(id=entry_id)
     topic = entry.topic
+
+    if request.user != topic.owner:
+        raise http.Http404("You are not allowed.")
 
     if request.method != 'POST':
         # Initial request; pre-fill form with the current entry.
